Refactor view forms and view registry
