--- src/components/Screens/__tests__/ViewForms.test.js
@@ -1,20 +1,0 @@
-/* eslint-env jest */
-
-import React from 'react';
-import { shallow } from 'enzyme';
-import { ViewForms } from '../ViewForms';
-
-const mockProps = {
-  forms: {},
-  nodes: {},
-  deleteForm: () => {},
-  openDialog: () => {},
-};
-
-describe('<ViewForms />', () => {
-  it('can render', () => {
-    const component = shallow(<ViewForms {...mockProps} />);
-
-    expect(component).toMatchSnapshot();
-  });
-});
