+++ src/components/Screens/screenIndex.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { get } from 'lodash';
 import StageEditorScreen from './StageEditorScreen';
 import SkipLogicEditorScreen from './SkipLogicEditorScreen';
-import ViewForms from './ViewForms';
+import ViewFormsScreen from './ViewFormsScreen';
 import FormEditorScreen from './FormEditorScreen';
-import Codebook from './VariableRegistry';
+import Codebook from './VariableRegistryScreen';
 import TypeEditorScreen from './TypeEditorScreen';
 import VariableEditorScreen from './VariableEditorScreen';
 
@@ -13,7 +13,7 @@ const NotFound = () => (<div> Screen not found </div>);
 const SCREEN_INDEX = {
   stage: StageEditorScreen,
   skip: SkipLogicEditorScreen,
-  forms: ViewForms,
+  forms: ViewFormsScreen,
   form: FormEditorScreen,
   codebook: Codebook,
   type: TypeEditorScreen,
